requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
dateparser==1.2.0
Django==4.2.9
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.2
fastapi-users==12.1.3
frozenlist==1.4.1
gevent==23.9.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.4
passlib==1.7.4
priority==2.0.0
psycopg2==2.9.9
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.27.1
websocket==0.2.1
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
